+++ appbuilder/tests/test_style_rewrite.py
@@ -35,14 +35,6 @@ class TestStyleRewriteComponent(unittest.TestCase):
         self.node = appbuilder.StyleRewrite(model=self.model_name)
         self.sc=StyleChoices.YINGXIAO
 
-    # def test_run_with_default_params(self):
-    #     """测试 run 方法使用默认参数"""
-    #     query = "帮我写一篇关于人体工学椅的文案"
-    #     msg = appbuilder.Message(query)
-    #     answer = self.node(msg)
-    #     self.assertIsNotNone(answer)
-        # 可以添加更多断言来检查 answer 的特定属性
-
     def test_to_chinese(self):
         result=self.sc.to_chinese()
         self.assertEqual(result,"营销话术")
@@ -55,30 +47,6 @@ class TestStyleRewriteComponent(unittest.TestCase):
         answer = self.node(msg, style=style)
         self.assertIsNotNone(answer)
         # 检查 answer 是否符合预期
-
-    # def test_run_with_invalid_params(self):
-    #     """测试 run 方法使用无效参数"""
-    #     query = "帮我写一篇关于人体工学椅的文案"
-    #     msg = appbuilder.Message(query)
-    #     style = "无效话术"
-    #     with self.assertRaises(ValueError):
-    #         self.node(msg, style=style)
-
-    # def test_run_with_different_style(self):
-    #     """测试不同的 style 参数值"""
-    #     node = appbuilder.StyleRewrite("ERNIE Speed-AppBuilder")
-    #     msg = appbuilder.Message("测试消息")
-    #     for style in StyleChoices:
-    #         with self.subTest(style=style):
-    #             answer = node(msg, style=style.value)
-    #             self.assertIsNotNone(answer)
-
-    # def test_run_with_stream_and_temperature(self):
-    #     """测试不同的 stream 和 temperature 参数值"""
-    #     node = appbuilder.StyleRewrite("ERNIE Speed-AppBuilder")
-    #     msg = appbuilder.Message("测试消息")
-    #     answer = node(msg, style="激励话术", stream=False, temperature=0.5)
-    #     self.assertIsNotNone(answer)
 
     def test_tool_eval_valid(self):
         """测试 tool_eval 方法使用有效参数"""
