+++ appbuilder/tests/test_style_writing.py
@@ -53,44 +53,6 @@ class TestStyleWritingComponent(unittest.TestCase):
         self.assertIsNotNone(answer)
         # 检查 answer 是否符合预期
 
-    # def test_run_with_invalid_params(self):
-    #     """测试 run 方法使用无效参数"""
-    #     query = "帮我写一篇关于人体工学椅的文案"
-    #     msg = appbuilder.Message(query)
-    #     style = "无效样式"
-    #     with self.assertRaises(ValueError):
-    #         self.node(msg, style_query=style)
-    #     time.sleep(1)
-
-    # def test_run_with_different_style_query(self):
-    #     """测试不同的 style_query 参数值"""
-    #     node = appbuilder.StyleWriting("ERNIE Speed-AppBuilder")
-    #     msg = appbuilder.Message("测试消息")
-    #     for style_query in StyleQueryChoices:
-    #         with self.subTest(style_query=style_query):
-    #             answer = node(msg, style_query=style_query.value, length=100)
-    #             self.assertIsNotNone(answer)
-    #         break
-    #     time.sleep(1)
-
-    # def test_run_with_different_length(self):
-    #     """测试不同的 length 参数值"""
-    #     node = appbuilder.StyleWriting("ERNIE Speed-AppBuilder")
-    #     msg = appbuilder.Message("测试消息")
-    #     for length in LengthChoices:
-    #         with self.subTest(length=length):
-    #             answer = node(msg, style_query="通用", length=length.value)
-    #             self.assertIsNotNone(answer)
-    #         break
-    #     time.sleep(1)
-
-    # def test_run_with_stream_and_temperature(self):
-    #     """测试不同的 stream 和 temperature 参数值"""
-    #     node = appbuilder.StyleWriting("ERNIE Speed-AppBuilder")
-    #     msg = appbuilder.Message("测试消息")
-    #     answer = node(msg, style_query="通用", length=100, stream=False, temperature=0.5)
-    #     self.assertIsNotNone(answer)
-    #     time.sleep(1)
 
     def test_tool_eval_valid(self):
         """测试 tool_eval 方法使用有效参数"""
